Don’t hardcode Windows platform version and arch in ScriptGenerator
`ScriptGenerator` was written at a time when we only built a single package
for both 32 and 64-bit Windows and as a result it had an Omnitruck url to
fetch package metadata where the platform architecture was hardcoded to
64-bit. This would cause consumers (such as `test-kitchen`) of
`ScriptGenerator`'s generated install script on 32-bit Windows to always
try to download and install a 64-bit Chef package. This commit removes the
hardcoded URL but a proper fix should probably be to deprecate
`Mixlib::Install::ScriptGenerator.new` in favor of only using
`Mixlib::Install.new`.

Signed-off-by: Scott Christopherson <scott@chef.io>