Fix Amazon 2 detection
The elif that looks for /etc/system-release was firing for Amazon 2 and the /etc/os-release parsing was never occurring. This moves all logic to the first check and then attempts to parse the data from /etc/os-release if the file exists (2015+). Slightly harder to grok due to the funky sh string check we have no, but it should be functional on all systems w/o risking impact to other RHEL systems.

Signed-off-by: Tim Smith <tsmith@chef.io>