Properly identify Windows 2019 in the install.ps1
Right now we're identifying Windows 2019 systems as being 2016. This isn't a critical issue as our 2016 and 2019 packages are actually the same thing, but it would be good to properly identify the OS we're on. At the very least this gets us better usage data for Windows releases. The unfortunate thing is that Microsoft hasn't bumped the NT major or minor versionns for 2019. Instead they just pumped the patch (build) of the release. 2016 is 14xxx and 2019 is 17xxx. Last match wins in these switch statements so if we see a 17xxx build then identify it as 2019. 18000 is Windows 10 so I don't see them releasing a 2019 build that is in the 18000 range. If they do for some reason we'll identify these as 2016, which would be fine. I think that's the overly cautious route to take unless someone has concerns.

Signed-off-by: Tim Smith <tsmith@chef.io>